Create AmountValidationDispatcher
Create a spearate module folder for validation dispatchers.
Create the first dispatcher, for amount.
Regroup global object.
