+++ app/js/main.js
@@ -4,10 +4,7 @@
  */
 module.exports = {
 	FormValidation: require( './lib/form_validation' ),
-	ReduxValidation: {
-		createValidationDispatcher: require( './lib/validation_dispatcher' ).createValidationDispatcher,
-		createValidationDispatcherCollection: require( './lib/validation_dispatcher_collection' ).createValidationDispatcherCollection
-	},
+	ValidationDispatchers: require( './lib/validation_dispatchers' ),
 	Components: require( './lib/form_components' ),
 	Store: require( './lib/store' ),
 	StoreUpdates: require( './lib/store_update_handling' ),
