+++ web/res/js/donationForm.js
@@ -44,13 +44,11 @@ $( function () {
 	WMDE.StoreUpdates.connectValidatorsToStore(
 		function ( initialValues ) {
 			return [
-				WMDE.ReduxValidation.createValidationDispatcher(
+				WMDE.ValidationDispatchers.createAmountValidationDispatcher(
 					WMDE.FormValidation.createAmountValidator( initData.data( 'validate-amount-url' ) ),
-					actions.newFinishPaymentDataValidationAction,
-					[ 'amount', 'paymentType' ],
 					initialValues
 				),
-				WMDE.ReduxValidation.createValidationDispatcher(
+				WMDE.ValidationDispatchers.createValidationDispatcher(
 					WMDE.FormValidation.createAddressValidator( 
 						initData.data( 'validate-address-url' ),
 						WMDE.FormValidation.DefaultRequiredFieldsForAddressType
@@ -59,13 +57,13 @@ $( function () {
 					[ 'addressType', 'salutation', 'title', 'firstName', 'lastName', 'companyName', 'street', 'postcode', 'city', 'country', 'email' ],
 					initialValues
 				),
-				WMDE.ReduxValidation.createValidationDispatcher(
+				WMDE.ValidationDispatchers.createValidationDispatcher(
 					WMDE.FormValidation.createEmailAddressValidator( initData.data( 'validate-email-address-url' ) ),
 					actions.newFinishEmailAddressValidationAction,
 					[ 'email' ],
 					initialValues
 				),
-				WMDE.ReduxValidation.createValidationDispatcher(
+				WMDE.ValidationDispatchers.createValidationDispatcher(
 					WMDE.FormValidation.createBankDataValidator(
 						initData.data( 'validate-iban-url' ),
 						initData.data( 'generate-iban-url' )
@@ -74,7 +72,7 @@ $( function () {
 					[ 'iban', 'bic', 'accountNumber', 'bankCode', 'debitType', 'paymentType' ],
 					initialValues
 				),
-				WMDE.ReduxValidation.createValidationDispatcher(
+				WMDE.ValidationDispatchers.createValidationDispatcher(
 					WMDE.FormValidation.createSepaConfirmationValidator(),
 					actions.newFinishSepaConfirmationValidationAction,
 					[ 'confirmSepa', 'confirmShortTerm' ],
